lambdas/analyzer/requirements.txt
boto3>=1.34.0
anthropic>=0.34.0
openai>=1.0.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langsmith>=0.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
